convert to min

+++ pipeline/0x01-apis/2-user_location.py
@@ -11,6 +11,6 @@ if __name__ == '__main__':
         print(resp.json()["location"])
     elif resp.status_code == 403:
         x = int(time.time()) - int(resp.headers['X-Ratelimit-Reset'])
-        print('Reset in {} min'.format(abs(x) / 60))
+        print('Reset in {} min'.format(int(abs(x) / 60)))
     else:
         print('Not found')
